Update km_mil_5.py
This is a basic list of code I wrote to see if I could perform some easy steps to calculate walking speed (distance/time) and show it in both kph and mph. The next step is to find ways to shorten the lines of code, if possible. I don't even think it's necessary to us try and except here but, as a first attempt, I wanted to cover all possible errors.